+++ backend/webserver/src/domain/task.rs
@@ -1,0 +1,9 @@
+use rocket::serde::{Deserialize, Serialize};
+
+
+#[derive(Deserialize)]
+#[derive(Serialize)]
+#[serde(crate = "rocket::serde")]
+pub struct Task {
+    pub code: String,
+}
