+++ backend/webserver/src/endpoints/index.rs
@@ -1,0 +1,6 @@
+
+
+#[get("/")]
+pub fn index() -> &'static str {
+    "Hello, world!"
+}
