+++ backend/webserver/src/endpoints/mod.rs
@@ -1,0 +1,5 @@
+
+pub mod haskell;
+pub mod index;
+pub mod cors;
+
