--- backend/webserver/src/writefile.rs
@@ -1,8 +1,0 @@
-use std::fs::File;
-use std::io::prelude::*;
-
-fn writeStruct(exercise : Task) -> std::io::Result<()>{    
-    let mut file = File::create("helloHaskell.hs")?;
-    file.write_all(exercise.code.as_bytes())?;
-    Ok(())
-}
